@@ -1,11 +1,13 @@
 module.exports = {
   extends: "eslint:recommended",
   parserOptions: {
-    ecmaVersion: 11
+    ecmaVersion: 11,
+    sourceType: "module"
   },
+  plugins: ["jest"],
   env: {
     node: true,
     es6: true,
-    mocha: true
+    "jest/globals": true 
   }
 }
